+++ src/components/Sample.vue
@@ -21,13 +21,14 @@
             >🔄</span
           >
         </div>
-        <div id="granularSliders" v-show="mode == 'granular'" class="space-out">
-          <div id="spray">Sp</div>
-          <div id="grainSize">Gs</div>
-          <div id="rate">Rt</div>
-          <div id="random">Rd</div>
+        <div class="granularSliders space-out" v-show="mode == 'granular'">
+          <!-- <div id="spray">Sp</div> -->
+          <div :id="`grainSize-${idx}`">Gs</div>
+          <div :id="`rate-${idx}`">Rt</div>
+          <div :id="`random-${idx}`">Rd</div>
         </div>
       </div>
+      <!-- show grains in DOM (instead of via <canvas>) -->
       <!-- <div v-show="mode === 'granular'" class="no-select" @click.stop>
         <div
           v-for="(grain, idx) in settings.granular.grains"
@@ -37,6 +38,17 @@
           ·
         </div>
       </div> -->
+      <div
+        id="controls"
+        v-show="controls === 'scale'"
+        class="no-select"
+        @click.stop
+      >
+        <div class="granularSliders space-out">
+          <div :id="`amplitude-${idx}`">Am</div>
+          <div :id="`timestretch-${idx}`">St</div>
+        </div>
+      </div>
       <div class="header" :id="`header-${idx}`">{{ name }}</div>
       <div id="waveform-wrapper">
         <div :class="`waveform-${idx}`" ref="waveformDiv"></div>
@@ -47,12 +59,8 @@
         <div id="mode-btn" @click.stop="toggleMode()">
           {{ mode === "sample" ? "▶️" : "*️⃣" }}
         </div>
-        <div
-          id="scale-btn"
-          @click.stop="toggleControls('scale')"
-          style="opacity: 0.5"
-        >
-          ↔️
+        <div id="scale-btn" @click.stop="toggleControls('scale')" style="">
+          ↕️
         </div>
       </div>
     </div>
@@ -61,7 +69,6 @@
 
 <script>
 import WaveSurfer from "wavesurfer.js"
-// import RegionsPlugin from "https://unpkg.com/wavesurfer.js@7/dist/plugins/regions.esm.js"
 import RegionsPlugin from "wavesurfer.js/dist/plugins/regions.esm.js"
 
 import Draggable from "gsap/Draggable"
@@ -91,13 +98,33 @@ export default {
         granular: {
           sourcePoint: null, // this is stored in [0, 1] range of the progress of the sample
           params: {
-            spray: { min: 0, max: 1, value: 1.0 }, // distance in function of grain size of sprea from source point
+            // spray: { min: 0, max: 1, value: 1.0 }, // distance in function of grain size of sprea from source point
             grainSize: { min: 0.05, max: 1.5, value: 1.5 },
-            rate: { min: 2.0, max: 0.1, value: 1.0 }, // how often a new grain is produced [2000ms, 100ms]
+            rate: { min: 0.1, max: 2, value: 1 }, // how often a new grain is produced [100ms, 2000ms]
             random: { min: 0.0, max: 1.0, value: 0.5 }, // how random (0 is at center, 1 is almost linearly, but still follows normal distribution)
           },
-          sliders: [],
+          sliders: {},
           grains: [],
+          envelope: {
+            attack: 0.2,
+            release: 0.2,
+          },
+          region: {
+            // store region between toggles of sample and granular mode
+            start: null,
+            end: null,
+          },
+        },
+        scale: {
+          params: {
+            amplitude: { min: 0, max: 1.5, value: 1 },
+            timestretch: {
+              min: mapNumber(100, 200 * 0.1, 200 * 2, 0.1, 2),
+              max: 2,
+              value: 1,
+            },
+          },
+          sliders: {},
         },
       },
       draggables: {
@@ -118,6 +145,20 @@ export default {
       if (this.distance > this.$root.reverbRadius) return 0
       return mapNumber(this.distance, 0, this.$root.reverbRadius, 1, 0)
     },
+
+    grainSize() {
+      return this.settings.granular.params.grainSize.value
+    },
+
+    invertedRate() {
+      return mapNumber(
+        this.settings.granular.params.rate.value,
+        this.settings.granular.params.rate.min,
+        this.settings.granular.params.rate.max,
+        this.settings.granular.params.rate.max,
+        this.settings.granular.params.rate.min
+      )
+    },
   },
 
   watch: {
@@ -135,7 +176,11 @@ export default {
         this.name,
         this.effectSendGainValue
       )
-      this.effectSendNode.gain.value = this.effectSendGainValue
+      // this.effectSendNode.gain.value = this.effectSendGainValue
+      this.effectSendNode.gain.exponentialRampToValueAtTime(
+        this.effectSendGainValue,
+        0.02
+      )
     },
 
     mode() {
@@ -149,6 +194,7 @@ export default {
     this.initWaveform()
     this.initAudio()
     this.initGranularSliders()
+    this.initScaleSliders()
     this.initCanvas()
 
     this.$root.$on("toggleControls", this.toggleControls)
@@ -170,8 +216,9 @@ export default {
         cursorColor: "transparent",
         cursorWidth: 1,
         barWidth: 2,
+        barHeight: 1, // goes from [0.1 to 4]
         barGap: 3,
-        normalize: true,
+        normalize: false, // if true, then barHeight for amplitude doesn't work
         interact: false,
         hideScrollbar: true,
         plugins: [RegionsPlugin.create({})],
@@ -232,6 +279,7 @@ export default {
       this.draggables = this.draggables[this.idx]
 
       this.wavesurfer = wavesurfer
+      window.wavesurfer = wavesurfer
       this.$root.wavesurfer = this.wavesurfer
     },
 
@@ -240,8 +288,13 @@ export default {
       this.audioNode = new Tone.Player(this.audio, () => {
         console.log("loaded audio", this.audio)
       })
-
+      this.audioNode.volume.value = -6
       window.audioNode = this.audioNode
+
+      // gain node to control via amplitude slider
+      this.audioGainNode = new Tone.Gain(
+        this.settings.scale.params.amplitude.value
+      )
 
       // init effect send (gain) node
       this.effectSendNode = new Tone.Gain(0)
@@ -249,9 +302,11 @@ export default {
       this.effectSendNode.connect(this.$root.reverbNode)
 
       // connect audio node to both effect send
-      this.audioNode.connect(this.effectSendNode)
-      // and destination
-      this.audioNode.connect(Tone.Master)
+      this.audioNode.connect(this.audioGainNode)
+      this.audioGainNode.connect(this.effectSendNode)
+      this.audioGainNode.connect(Tone.Master)
+
+      // this.audioNode.chain(this.effectSendNode, Tone.Master)
     },
 
     initGranularSliders() {
@@ -260,20 +315,8 @@ export default {
         mode: "relative", // 'relative' or 'absolute'
       }
 
-      // spray
-      const spraySlider = new Nexus.Slider("#spray", {
-        ...opts,
-        ...this.settings.granular.params.spray,
-      })
-      this.settings.granular.sliders.spray = spraySlider
-      spraySlider.on("change", (val) => {
-        console.log("spray is now", val)
-        this.settings.granular.params.spray.value = val
-        this.initRegions()
-      })
-
       // grain size
-      const grainSizeSlider = new Nexus.Slider("#grainSize", {
+      const grainSizeSlider = new Nexus.Slider(`#grainSize-${this.idx}`, {
         ...opts,
         ...this.settings.granular.params.grainSize,
       })
@@ -281,23 +324,22 @@ export default {
       grainSizeSlider.on("change", (val) => {
         console.log("grainSize is now", val)
         this.settings.granular.params.grainSize.value = val
-        this.initRegions()
       })
 
       // rate
-      const rateSlider = new Nexus.Slider("#rate", {
+      const rateSlider = new Nexus.Slider(`#rate-${this.idx}`, {
         ...opts,
-        ...this.settings.granular.params.rateSlider,
+        ...this.settings.granular.params.rate,
       })
       this.settings.granular.sliders.rate = rateSlider
       rateSlider.on("change", (val) => {
-        console.log("rate is now", val)
+        console.log("rate is now", val, this.invertedRate)
         this.settings.granular.params.rate.value = val
         this.updateRateInterval()
       })
 
       // random
-      const randomSlider = new Nexus.Slider("#random", {
+      const randomSlider = new Nexus.Slider(`#random-${this.idx}`, {
         ...opts,
         ...this.settings.granular.params.randomSlider,
       })
@@ -308,14 +350,71 @@ export default {
       })
 
       // colors
-      spraySlider.colorize("accent", "var(--blue)")
-      spraySlider.colorize("fill", "var(--blue-light)")
       grainSizeSlider.colorize("accent", "var(--blue)")
       grainSizeSlider.colorize("fill", "var(--blue-light)")
       rateSlider.colorize("accent", "var(--blue)")
       rateSlider.colorize("fill", "var(--blue-light)")
       randomSlider.colorize("accent", "var(--blue)")
       randomSlider.colorize("fill", "var(--blue-light)")
+    },
+
+    initScaleSliders() {
+      const opts = {
+        size: [20, 100],
+        mode: "relative", // 'relative' or 'absolute'
+      }
+
+      // amplitude
+      const amplitudeSlider = new Nexus.Slider(`#amplitude-${this.idx}`, {
+        ...opts,
+        ...this.settings.scale.params.amplitude,
+      })
+      this.settings.scale.sliders.amplitude = amplitudeSlider
+      amplitudeSlider.on("change", (val) => {
+        console.log("amplitude is now", val)
+        const mappedBarHeight = mapNumber(
+          val,
+          0,
+          this.settings.scale.params.amplitude.max,
+          0,
+          this.settings.scale.params.amplitude.max
+        )
+        this.wavesurfer.setOptions({ barHeight: mappedBarHeight })
+        console.log("using amplitude of ", val, "to apply in gain node")
+        // this.audioGainNode.gain.value = val
+        this.audioGainNode.gain.exponentialRampToValueAtTime(val, 0.02)
+        this.settings.scale.params.amplitude.value = val
+      })
+
+      // timestretch
+      const timestretch = new Nexus.Slider(`#timestretch-${this.idx}`, {
+        ...opts,
+        ...this.settings.scale.params.timestretch,
+      })
+      this.settings.scale.timestretch = timestretch
+      timestretch.on("change", (val) => {
+        console.log("timestretch is now", val)
+        const container = document.querySelector(".container")
+        container.style.width = `${this.width * val}px`
+        console.log(this.width, container.style.width)
+        this.settings.scale.params.timestretch.value = val
+        const invertedPlaybackRate = mapNumber(
+          val,
+          this.settings.scale.params.timestretch.min,
+          this.settings.scale.params.timestretch.max,
+          this.settings.scale.params.timestretch.max,
+          this.settings.scale.params.timestretch.min
+        )
+        console.log({ invertedPlaybackRate })
+        this.audioNode.playbackRate = invertedPlaybackRate
+        this.wavesurfer.setOptions({ audioRate: invertedPlaybackRate })
+      })
+
+      // colors
+      amplitudeSlider.colorize("accent", "var(--blue)")
+      amplitudeSlider.colorize("fill", "var(--blue-light)")
+      timestretch.colorize("accent", "var(--blue)")
+      timestretch.colorize("fill", "var(--blue-light)")
     },
 
     initCanvas() {
@@ -335,6 +434,9 @@ export default {
     },
 
     play() {
+      // this.audioNode.seek(0)
+      this.wavesurfer.seekTo(0)
+      this.audioNode.restart()
       this.wavesurfer.play()
       this.isPlaying = true
       this.audioNode.start()
@@ -362,81 +464,77 @@ export default {
         // this.wavesurfer.setOptions({ cursorColor: "black" })
         this.wavesurfer.setOptions({ progressColor: "lightgray" })
         this.wavesurfer.setOptions({ interact: "true" })
-        this.initRegions()
+        this.initRegion()
       } else {
         // this.wavesurfer.setOptions({ cursorColor: "transparent" })
         this.wavesurfer.setOptions({ interact: "false" })
+        this.wavesurfer.setOptions({ progressColor: "lightblack" })
         clearInterval(this.settings.granular.params.rate.interval)
         this.clearCanvas()
+        // store regions before leaving granular mode
+        const region = this.regions.regions[0]
+        this.settings.granular.region = {
+          start: region.start,
+          end: region.end,
+        }
         this.regions.clearRegions()
       }
     },
 
     updateRateInterval() {
       clearInterval(this.settings.granular.params.rate.interval)
+      this.addGrain()
       this.settings.granular.params.rate.interval = setInterval(() => {
         this.addGrain()
-      }, this.settings.granular.params.rate.value * 1000)
+      }, this.invertedRate * 1000)
     },
 
-    initRegions() {
-      const relX =
-        this.settings.granular.sourcePoint * this.audioNode.buffer.duration
-      const randOffsetRange =
-        this.settings.granular.params.spray.value *
-        this.settings.granular.params.grainSize.value
-      const start = relX - randOffsetRange
-      const end = relX + randOffsetRange
-      this.settings.granular.params.grainSize.value
+    initRegion() {
+      let start, end
+      if (!this.settings.granular.region.start) {
+        const relX =
+          this.settings.granular.sourcePoint * this.audioNode.buffer.duration
+        const randOffsetRange = this.grainSize * 2
+        start = relX - randOffsetRange
+        end = relX + randOffsetRange
+      } else {
+        start = this.settings.granular.region.start
+        end = this.settings.granular.region.end
+      }
       this.regions.clearRegions()
       this.regions.addRegion({
         start,
         end,
         content: "",
         color: "rgba(170, 197, 216, 0.2)",
-        resize: false,
+        resize: true,
       })
+
+      window.regions = this.regions
 
       this.regions.on("region-updated", (region) => {
         console.log("Updated region", region)
-        const startRelX = mapNumber(
-          region.start,
+        const middleX = (region.end - region.start) / 2 + region.start
+        const sourcePoint = mapNumber(
+          middleX,
           0,
           this.audioNode.buffer.duration,
           0,
           1
         )
-        this.settings.granular.sourcePoint = startRelX
-      })
-
-      this.regions.enableDragSelection({
-        color: "rgba(255, 0, 0, 0.1)",
+        this.settings.granular.sourcePoint = sourcePoint
       })
 
       this.updateRateInterval()
     },
 
     addGrain() {
-      console.log("adding grain", this.settings.granular.grains)
-
       // access buffer
       window.audioNode = this.audioNode
       this.buffer = this.audioNode.buffer.getChannelData(0)
       window.buffer = this.buffer
-      // this.buffer = null // TODO
 
-      // to create a buffer, one needs to clone and then change the data
-      // const clonedBuffer = Tone.context.createBuffer(
-      //   1,
-      //   this.buffer.length,
-      //   44100
-      // )
-      // const bufferData = clonedBuffer.getChannelData(0) // TODO: just using one channel for now (even though there are two channels)
-      // for (let i = 0; i < this.buffer.length; i++) {
-      //   bufferData[i] = this.audioNode.buffer.getChannelData(0)[i]
-      // }
-
-      // // determine where to read the grain from
+      // determine where to read the grain from
       const baseOffset = mapNumber(
         this.settings.granular.sourcePoint,
         0,
@@ -448,47 +546,79 @@ export default {
       // uniform random
       // const randOffsetRange =
       //   this.settings.granular.params.spray.value *
-      //   this.settings.granular.params.grainSize.value
+      //   this.grainSize
       // const randOffset = randomFloat(0, randOffsetRange) - randOffsetRange / 2
-      const randOffset = randomGaussian(
+
+      const region = this.regions.regions[0]
+      const space = (region.end - region.start) / 2
+      console.log({ space })
+      const mappedStdDev = mapNumber(
+        this.settings.granular.params.random.value,
         0,
-        this.settings.granular.params.random.value
+        1,
+        0,
+        space
       )
+      const randOffset = randomGaussian(0, mappedStdDev)
+
+      console.log(randOffset)
 
       const grainOffset = clamp(
         baseOffset + randOffset,
-        0,
-        this.audioNode.buffer.duration
+        region.start,
+        region.end
       )
 
       // create a grain buffer
-      // const grain = Tone.context.createBufferSource()
-      // grain.buffer = clonedBuffer
       const grain = new Tone.Player(this.audioNode.buffer)
-      // grain.addEventListener("ended", () => {
       setTimeout(() => {
-        this.settings.granular.grains.pop()
-      }, 1000)
+        this.settings.granular.grains.shift()
+        this.drawGrains()
+      }, this.grainSize * 1000)
 
-      // connect audio node to both effect send
-      grain.connect(this.effectSendNode)
-      // and destination
-      grain.connect(Tone.Master)
-
+      grain.volume.value = -6
       window.grain = grain
 
-      window.grain = grain
+      // console.log({ grainOffset })
+
+      // set offsets according to envelope
+      const now = Tone.context.currentTime
+      const gs = this.grainSize
+      const useEnvelope = true
+      if (!useEnvelope) {
+        // connect audio node to both effect send
+        grain.connect(this.effectSendNode)
+        // and destination
+        grain.connect(Tone.Master)
+      } else {
+        const gain = new Tone.Gain(0)
+        grain.connect(gain)
+        gain.connect(this.effectSendNode)
+        gain.connect(Tone.Master)
+
+        const attackOffset = now + this.settings.granular.envelope.attack
+        const releaseOffset = Math.max(
+          gs - this.settings.granular.envelope.release,
+          this.settings.granular.envelope.release
+        )
+
+        gain.gain.linearRampToValueAtTime(1, attackOffset)
+        gain.gain.linearRampToValueAtTime(0, releaseOffset)
+        // gain.gain.linearRampToValueAtTime(1, "+0.02")
+
+        // console.log({
+        //   now,
+        //   attackOffset,
+        //   releaseOffset,
+        //   grainSize: gs,
+        // })
+      }
 
       // play grain
-      console.log({ grainOffset })
-      grain.start(
-        Tone.context.currentTime,
-        grainOffset,
-        this.settings.granular.params.grainSize.value
-      )
+      grain.start(now, grainOffset, this.grainSize)
 
-      // draw grain
-      console.log({ grainOffset })
+      // draw grains
+      // console.log({ grainOffset })
       const x = mapNumber(
         grainOffset,
         0,
@@ -515,9 +645,9 @@ export default {
 
       // drawing source point
       // const r = 5
-      // this.canvasCtx.arc(this.x, this.height / 2, r, 0, 2 * Math.PI)
       // this.canvasCtx.stroke()
-      this.canvasCtx.fill()
+      // this.canvasCtx.fillStyle = "red"
+      // this.canvasCtx.arc(this.x, this.height / 2, r, 0, 2 * Math.PI)
 
       for (let i = 0; i < this.settings.granular.grains.length; i++) {
         const { x } = this.settings.granular.grains[i]
@@ -527,7 +657,7 @@ export default {
         // const minW = 5
         // const w =
         //   (this.width / this.audioNode.buffer.duration) *
-        //   this.settings.granular.params.grainSize.value
+        //   this.grainSize
         // const h = 20
         // const size = Math.max(minW, w)
         // this.canvasCtx.fillRect(x - w / 2, this.height / 2 - h / 2, size, h)
@@ -535,9 +665,7 @@ export default {
         // circle
         this.canvasCtx.beginPath()
         const r =
-          ((this.width / this.audioNode.buffer.duration) *
-            this.settings.granular.params.grainSize.value) /
-          2
+          ((this.width / this.audioNode.buffer.duration) * this.grainSize) / 2
         this.canvasCtx.arc(x, this.height / 2, r, 0, 2 * Math.PI)
         // this.canvasCtx.stroke()
         this.canvasCtx.fill()
@@ -635,7 +763,7 @@ export default {
   width: 100%;
 }
 
-#granularSliders {
+.granularSliders {
   display: flex;
   pointer-events: auto; // to override the none of the parent
   width: 100%;